Remove 'subject' and 'its' syntax.
Our rspec specification will initially only include basic rspec spec definition DSL of it "does something" for simplicity's sake. There's a lot of shorthand that is very clean and concise, but it can be confusing during the adoption phase. As we learn what we do and don't like about writing out tests longhand, we'll add recommendations for shorthand usage as needed.